+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Error when loading a previously run `Batch` or `ComponentModeler` containing custom data.
+- Error when plotting mode plane PML and the simulation has symmetry.
 
 ## [2.7.1]